Obstacle Discovered - Directory Contents
Currently, directories are stored as arrays inside a base array with
the key being the full path.  These arrays also have a 'Contents' key
which was planned to point by reference to the contents of the
directory.  The advantage of this is the simplicity of not having to
maintain a full tree (all parents).

The obsticle is that right now the MockDirectoryPointer uses the
FileSystemInterface to access the MockFileSystem.  While typing this,
it has become obvious to me that my mistake was using the
FileSystemInterface as the type of the constructor for the
MockDirectoryPointer.  The Mock elements should not ever use the real
file system.

Therefore, the next step would be to create a MockFileSystemInterface
that extends FileSystemInterface and has the additional needed
functions.